requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
auto_gptq==0.7.1
Automat==24.8.1
base58==2.1.1
blinker==1.8.2
blis==0.7.11
boto3==1.38.32
botocore==1.38.32
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
conllu==4.4.2
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
cymem==2.0.11
datasets==2.10.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.6
docker==7.1.0
docker-pycreds==0.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
exceptiongroup==1.2.2
fairscale==0.4.6
filelock==3.13.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.3.1
gekko==1.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.14.0
h2==4.1.0
h5py==3.14.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.32.4
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
lmdb==1.6.2
lxml==5.3.0
marisa-trie==1.2.1
MarkupSafe==2.1.5
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.14
murmurhash==1.0.13
neo4j==5.28.1
networkx==3.3
nltk==3.9.1
numpy==2.1.2
optimum==1.25.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
parsel==1.9.1
pathlib_abc==0.1.1
pathtools==0.1.2
pathy==0.11.0
peft==0.15.2
pillow==11.0.0
pluggy==1.6.0
preshed==3.0.10
promise==2.3
propcache==0.2.0
Protego==0.3.1
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
py-rouge==1.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.8.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==24.2.1
PySocks==1.7.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32==308
PyYAML==6.0.2
queuelib==1.7.0
realtime==2.0.6
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
responses==0.18.0
retrying==1.3.4
rich==12.6.0
rouge==1.0.1
rsa==4.9.1
s3transfer==0.13.0
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
scrapinghub==2.4.0
Scrapy==2.11.2
scrapy-selenium==0.0.7
selenium==4.32.0
sentencepiece==0.2.0
sentry-sdk==2.29.1
service-identity==24.1.0
setproctitle==1.3.6
shortuuid==1.0.13
six==1.16.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
storage3==0.8.2
StrEnum==0.4.15
supabase==2.9.1
supafunc==0.6.2
sympy==1.13.1
tensorboardX==2.6.2.2
termcolor==1.1.0
thinc==8.0.17
threadpoolctl==3.6.0
tldextract==5.1.2
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.7.0
typer==0.4.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
w3lib==2.2.1
wandb==0.12.21
wasabi==0.10.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
word2number==1.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.16.0
zope.interface==7.1.0
